feat: StudyListTable Refactor
