+++ platform/ui/src/components/StudyListExpandedRow/index.js
@@ -1,0 +1,1 @@
+export { default as StudyListExpandedRow } from './StudyListExpandedRow';
