+++ platform/ui/src/mocks/studyList.json
@@ -4,7 +4,7 @@
       "StudyInstanceUID": "1.2.840.113619.2.5.1762583153.215519.978957063.78",
       "StudyDescription": "BRAIN SELLA",
       "AccessionNumber": "11788761116031",
-      "StudyDate": "20010108",
+      "StudyDate": "2020-03-26T23:33:59.073Z",
       "StudyTime": "120022",
       "PatientName": "MISTER^MR",
       "PatientId": "832040",
